finish refactoring, init tests
