+++ jhotdraw-samples/jhotdraw-samples-misc/src/main/java/org/jhotdraw/samples/svg/bridge/RectImageBridge.java
@@ -1,4 +1,4 @@
-package org.jhotdraw.samples.bridge;
+package org.jhotdraw.samples.svg.bridge;
 
 import org.jhotdraw.samples.SPI.RectImage;
